% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{sir_model}
\alias{sir_model}
\title{Example SIR model}
\usage{
sir_model(S0 = 999, I0 = 1, R0 = 0, beta = 1, gamma = 0.5, time = 100)
}
\arguments{
\item{S0, I0, R0}{Initial number of individuals in the S, I, and R states.}

\item{beta, gamma}{Transmission rate and recovery rate, respectively.}

\item{time}{Duration of the simulation.}
}
\value{
An object of class \code{c("ode_model", "modeller")}. Use
\code{\link[=run_model]{run_model()}} to simulate, \link[=plot.model_result]{plot()} to visualise,
and \code{\link[=show_model]{show_model()}} to launch the interactive app.
}
\description{
Returns an 'example' SIR ODE model for testing purposes. There is a single
population with homogeneous mixing, a transmission rate \code{beta} and a
recovery rate \code{gamma} in the model.
}
\examples{
m = sir_model()
result = run_model(m)
plot(result)

}
